+++ CHANGES.rst
@@ -5,9 +5,12 @@ Changelog
 3.1.3 (unreleased)
 ------------------
 
+- Handle text-larger class in slate2html.
+  [cekk]
+- Do not pretty slate2html output to avoid not needed spaces between tags.
+  [cekk]
 - Changed label from Parametro to Dato.
   [eikichi18]
-
 
 3.1.2 (2024-12-05)
 ------------------